requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
httpx==0.28.1
openai==1.68.0
python-telegram-bot==21.7
